% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiattr.R
\name{build_target_scores}
\alias{build_target_scores}
\title{build one score table per declared column}
\usage{
build_target_scores(dat_raw_anon, targets, row_number, split, fn_name)
}
\arguments{
\item{dat_raw_anon}{dataframe of raw_anon form}

\item{targets}{an already-expanded named character vector}

\item{row_number}{name of the row-number column *before* the RAW_/ANON_
prefixing done by [join_raw_anon_data()] (default: "ROW_NUMBER")}

\item{split}{separator passed to [score_dist()] for `"dist"` columns}

\item{fn_name}{name used in error messages}
}
\value{
a named list of "reid_scores" tables, in the order of `targets`
}
\description{
build one score table per declared column
}
\keyword{internal}
