% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsim.R
\name{check_tversky_weights}
\alias{check_tversky_weights}
\title{check the Tversky weights}
\usage{
check_tversky_weights(alpha, beta, fn_name)
}
\arguments{
\item{alpha, beta}{the weights to check}

\item{fn_name}{calling function, for the error message}
}
\value{
`invisible(TRUE)`; stops otherwise
}
\description{
check the Tversky weights
}
\keyword{internal}
