% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\encoding{UTF-8}
\name{create_dummy_transaction_data}
\alias{create_dummy_transaction_data}
\title{create dummy transaction data}
\usage{
create_dummy_transaction_data(people = 100, size = 2)
}
\arguments{
\item{people}{number of people}

\item{size}{mean record number}
}
\value{
a tibble with columns ROW_NUMBER, ID, NUM_STATIC, NUM_DYNAMIC,
  BIN, CHAR containing `people * size` rows of randomly generated dummy
  transaction data.
}
\description{
create dummy transaction data
}
\examples{
data_tran = create_dummy_transaction_data(people = 10, size = 4)
}
