% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{descendants_of}
\alias{descendants_of}
\title{look up the descendants of a node}
\usage{
descendants_of(hierarchy, attribute, node)
}
\arguments{
\item{hierarchy}{a "reid_hierarchy" object, or NULL}

\item{attribute}{attribute name}

\item{node}{node name}
}
\value{
a character vector of the nodes strictly below `node`
}
\description{
look up the descendants of a node
}
\keyword{internal}
