% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{dummy_qi_knowledge}
\alias{dummy_qi_knowledge}
\title{the attacker knowledge specification that matches create_dummy_qi_data()}
\usage{
dummy_qi_knowledge(level = c("W", "M", "S"), ...)
}
\arguments{
\item{level}{one of "W", "M", "S"}

\item{...}{further arguments passed to [attacker_knowledge()]}
}
\value{
an [attacker_knowledge()] object for the columns
  [create_dummy_qi_data()] generates
}
\description{
the attacker knowledge specification that matches create_dummy_qi_data()
}
\examples{
dummy_qi_knowledge("W")

}
