% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{gen_escape_regex}
\alias{gen_escape_regex}
\title{escape a literal string for use inside a regular expression}
\usage{
gen_escape_regex(v)
}
\arguments{
\item{v}{character vector}
}
\value{
`v` with the regex metacharacters backslash-escaped
}
\description{
escape a literal string for use inside a regular expression
}
\keyword{internal}
