% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{gen_format_number}
\alias{gen_format_number}
\title{render a number without scientific notation}
\usage{
gen_format_number(x)
}
\arguments{
\item{x}{a single numeric value}
}
\value{
a string
}
\description{
`format()` would print 1e8 as "1e+08" and round 1234567.8 to seven
significant digits; either would be re-read as the wrong endpoint.
}
\keyword{internal}
