% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{gen_normalize}
\alias{gen_normalize}
\title{map fullwidth characters onto their ASCII equivalents}
\usage{
gen_normalize(v)
}
\arguments{
\item{v}{a single string}
}
\value{
the string with fullwidth digits and punctuation replaced
}
\description{
map fullwidth characters onto their ASCII equivalents
}
\keyword{internal}
