% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{is_generalization_mask}
\alias{is_generalization_mask}
\title{is this value a suppression mask?}
\usage{
is_generalization_mask(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
is this value a suppression mask?
}
\keyword{internal}
