% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{is_generalization_wildcard}
\alias{is_generalization_wildcard}
\title{does this value mean "suppressed / any value"?}
\usage{
is_generalization_wildcard(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
does this value mean "suppressed / any value"?
}
\keyword{internal}
