% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{join_row_anon_data}
\alias{join_row_anon_data}
\title{再識別アルゴリズム適用に便利な形のデータを作る}
\usage{
join_row_anon_data(raw, anon, raw_header = "RAW_",
  anon_header = "ANON_")
}
\arguments{
\item{raw}{匿名加工前データ(data.frame}

\item{anon}{匿名加工後データ(data.frame)}

\item{raw_header}{加工前データの列の先頭に着けるラベル(デフォルトはRAW_)}

\item{anon_header}{加工後データの列の先頭に着けるラベル(デフォルトはAN_)}
}
\description{
再識別アルゴリズム適用に便利な形のデータを作る
}
