% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiattr.R
\name{normalize_one_score}
\alias{normalize_one_score}
\title{rescale a single score table}
\usage{
normalize_one_score(s, method, arg = "scores")
}
\arguments{
\item{s}{a score table}

\item{method}{normalisation method, already matched}

\item{arg}{argument name used in error messages}
}
\value{
a "reid_scores" table
}
\description{
rescale a single score table
}
\keyword{internal}
