% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiattr.R
\name{one_axis_informativeness}
\alias{one_axis_informativeness}
\title{test one axis for information about identity}
\usage{
one_axis_informativeness(s, axis, alpha, arg = "scores")
}
\arguments{
\item{s}{a score table}

\item{axis}{name to report the axis under}

\item{alpha}{significance level}

\item{arg}{argument name used in error messages}
}
\value{
a one-row data frame
}
\description{
test one axis for information about identity
}
\keyword{internal}
