% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal.R
\name{one_k_unicity}
\alias{one_k_unicity}
\title{unicity at one k and one resolution pair}
\usage{
one_k_unicity(
  traces,
  by_point,
  this_k,
  n_samples,
  n_individuals,
  n_points,
  tr,
  sr
)
}
\arguments{
\item{traces}{list of distinct point codes per individual}

\item{by_point}{list of individual codes per point}

\item{this_k}{number of known points}

\item{n_samples}{maximum subsets drawn per individual}

\item{n_individuals}{number of individuals}

\item{n_points}{number of distinct points}

\item{tr, sr}{the resolutions, copied into the result row}
}
\value{
a one-row data frame
}
\description{
unicity at one k and one resolution pair
}
\keyword{internal}
