% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{print.attacker_knowledge}
\alias{print.attacker_knowledge}
\title{print an attacker knowledge model}
\usage{
\method{print}{attacker_knowledge}(x, ...)
}
\arguments{
\item{x}{an "attacker_knowledge" object}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
print an attacker knowledge model
}
