% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiattr.R
\name{print.reid_axis_report}
\alias{print.reid_axis_report}
\title{print an axis informativeness report}
\usage{
\method{print}{reid_axis_report}(x, ...)
}
\arguments{
\item{x}{a "reid_axis_report" data frame}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
print an axis informativeness report
}
