% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{print.reid_blocking}
\alias{print.reid_blocking}
\title{print a blocking record}
\usage{
\method{print}{reid_blocking}(x, ...)
}
\arguments{
\item{x}{a "reid_blocking" object}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
print a blocking record
}
