% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.reid_evaluation}
\alias{print.reid_evaluation}
\title{print a reidentification evaluation}
\usage{
\method{print}{reid_evaluation}(x, ...)
}
\arguments{
\item{x}{a "reid_evaluation" object}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
print a reidentification evaluation
}
