% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{print.reid_hierarchy}
\alias{print.reid_hierarchy}
\title{print a generalisation hierarchy}
\usage{
\method{print}{reid_hierarchy}(x, ...)
}
\arguments{
\item{x}{a "reid_hierarchy" object}

\item{...}{unused}
}
\value{
`x`, invisibly
}
\description{
print a generalisation hierarchy
}
