% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{print.reid_scores}
\alias{print.reid_scores}
\title{print a score table}
\usage{
\method{print}{reid_scores}(x, ...)
}
\arguments{
\item{x}{a "reid_scores" object}

\item{...}{passed on to the data frame print method}
}
\value{
`x`, invisibly
}
\description{
print a score table
}
