% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{reid_split_score_types}
\alias{reid_split_score_types}
\title{the score types whose function takes a `split` separator}
\usage{
reid_split_score_types()
}
\value{
character vector of score type names
}
\description{
the score types whose function takes a `split` separator
}
\keyword{internal}
