% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{score_fn_for_type}
\alias{score_fn_for_type}
\title{pick the score function for a declared score type}
\usage{
score_fn_for_type(type)
}
\arguments{
\item{type}{one of the values [reid_score_types()] lists}
}
\value{
the corresponding `score_*()` function
}
\description{
pick the score function for a declared score type
}
\keyword{internal}
