% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{validate_tie_tolerance}
\alias{validate_tie_tolerance}
\title{validate a tie tolerance}
\usage{
validate_tie_tolerance(tolerance, fn_name)
}
\arguments{
\item{tolerance}{the value to check}

\item{fn_name}{calling function, for the message}
}
\value{
`tolerance`, invisibly
}
\description{
validate a tie tolerance
}
\keyword{internal}
