% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{calc_KL}
\alias{calc_KL}
\title{calc KL divergence from 2 character vector which has distribution expression (A:B:C:...)}
\usage{
calc_KL(x, y, split = ":")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{split}{split (default: ":")}
}
\value{
numeric scalar, the KL divergence between the normalized
  distributions parsed from `x` and `y`.
}
\description{
calc KL divergence from 2 character vector which has distribution expression (A:B:C:...)
}
\keyword{internal}
