% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{create_dummy_master_data}
\alias{create_dummy_master_data}
\title{再識別用のダミーマスターデータ生成}
\usage{
create_dummy_master_data(people = 100)
}
\arguments{
\item{people}{人数}
}
\description{
再識別用のダミーマスターデータ生成
}
