% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\encoding{UTF-8}
\name{create_dummy_master_data}
\alias{create_dummy_master_data}
\title{create dummy master data}
\usage{
create_dummy_master_data(people = 100)
}
\arguments{
\item{people}{number of people}
}
\description{
create dummy master data
}
