% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\name{create_dummy_transaction_data}
\alias{create_dummy_transaction_data}
\title{再識別用のダミーマスターデータ生成}
\usage{
create_dummy_transaction_data(people = 100, size = 2)
}
\arguments{
\item{people}{人数}

\item{size}{平均トランザクション数}
}
\description{
再識別用のダミーマスターデータ生成
}
\examples{
data_tran = create_dummy_transaction_data(people = 10, size = 4)
}
