% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{distribution_distance}
\alias{distribution_distance}
\title{calculate distribution distance (by using L2 norm) from 2 character vector which has distribution expression (A:B:C:...)}
\usage{
distribution_distance(x, y, split = ":")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{split}{split (default: ":")}
}
\description{
calculate distribution distance (by using L2 norm) from 2 character vector which has distribution expression (A:B:C:...)
}
