% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data.R
\encoding{UTF-8}
\name{join_raw_anon_data}
\alias{join_raw_anon_data}
\title{create reid-format data from raw and anon data frame}
\usage{
join_raw_anon_data(raw, anon, raw_header = "RAW_",
  anon_header = "ANON_")
}
\arguments{
\item{raw}{raw data frame}

\item{anon}{anonymized data frame}

\item{raw_header}{strings which is added for columns from raw data}

\item{anon_header}{strings which is added for columns from anon data}
}
\description{
create reid-format data from raw and anon data frame
}
