% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{print.reid_stability}
\alias{print.reid_stability}
\title{print a reid_stability summary}
\usage{
\method{print}{reid_stability}(x, ...)
}
\arguments{
\item{x}{a "reid_stability" object}

\item{...}{ignored}
}
\value{
`x`, invisibly
}
\description{
print a reid_stability summary
}
