% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{reid_by_char}
\alias{reid_by_char}
\title{reidentify by character static column}
\usage{
reid_by_char(dat_raw_anon, target, row_number = "ROW_NUMBER")
}
\arguments{
\item{dat_raw_anon}{dataframe of raw_anon form}

\item{target}{target column}

\item{row_number}{row number column name(default: "ROW_NUMBER")}
}
\description{
reidentify by character static column
}
