% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{reid_by_dist}
\alias{reid_by_dist}
\title{reidentify by single list column (list A, B, C is expressed by "A:B:C")}
\usage{
reid_by_dist(dat_raw_anon, target, row_number = "ROW_NUMBER",
  split = ":")
}
\arguments{
\item{dat_raw_anon}{dataframe of raw_anon form}

\item{target}{target column}

\item{row_number}{row number column name(default: "ROW_NUMBER")}

\item{split}{character for split _DIST value (default: ":")}
}
\description{
reidentify by single list column (list A, B, C is expressed by "A:B:C")
}
