% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify.R
\name{reid_by_num}
\alias{reid_by_num}
\title{reidentify by single num static column by using L2 norm}
\usage{
reid_by_num(dat_raw_anon, target, row_number = "ROW_NUMBER")
}
\arguments{
\item{dat_raw_anon}{dataframe of raw_anon form}

\item{target}{target column}

\item{row_number}{row number column name(default: "ROW_NUMBER")}
}
\description{
reidentify by single num static column by using L2 norm
}
