% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify-package.R
\docType{package}
\name{reidentify-package}
\alias{reidentify}
\alias{reidentify-package}
\title{reidentify: Reidentify and check the safety of Anonymized data}
\description{
Check the safety of Anonymized data by estimating reid-attack success rate.
}
\keyword{internal}
