% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify-package.R
\docType{package}
\name{reidentify-package}
\alias{reidentify}
\alias{reidentify-package}
\title{reidentify: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
