% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_transaction_to_master.R
\name{transform_transaction_to_master}
\alias{transform_transaction_to_master}
\title{transform transaction -> master}
\usage{
transform_transaction_to_master(dat, ID = "ID", collapse = ":")
}
\arguments{
\item{dat}{transaction data frame}

\item{ID}{identifier name}

\item{collapse}{separation character (defalt ":")}
}
\description{
transform transaction -> master
}
