% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_transaction_to_master.R
\name{transform_transaction_to_master}
\alias{transform_transaction_to_master}
\title{トランザクションデータを無理やりマスタの形にする}
\usage{
transform_transaction_to_master(dat, ID = "ID", collapse = ":")
}
\arguments{
\item{dat}{トランザクションデータ}

\item{ID}{識別子の列名}

\item{collapse}{圧縮時の区切り文字}
}
\description{
トランザクションデータを無理やりマスタの形にする
}
